\name{iAF1260}
\docType{data}
\alias{iAF1260}

\encoding{utf8}

\title{Escherichia coli Metabolic Model iAF1260}

\description{
  The dataset is a genome scale metabolic network of the \emph{E. coli}. It consists of 2077 internal reactions, 304 exchange
  reactions and a biomass objective function.
}

\usage{data(iAF1260)}

\format{An object of class \code{modelorg}}

\references{
Feist AM, Henry CS, Reed JL, Krummenacker M, Joyce AR, Karp PD,
Broadbelt LJ, Hatzimanikatis V, Palsson BØ (2007) A genome-scale
metabolic reconstruction for Escherichia coli K-12 MG1655
that accounts for 1260 ORFs and thermodynamic information.
Mol Syst Biol \emph{3}: 121}

\keyword{datasets}
\keyword{iAF1260}