\name{iJO1366}
\docType{data}
\alias{iJO1366}

\encoding{utf8}

\title{Escherichia coli Metabolic Model iJO1366}

\description{
  The dataset is a genome scale metabolic network of the \emph{E. coli}. It consists of 2253 internal reactions, 330 exchange
  reactions and a biomass objective function.
}

\usage{data(iJO1366)}

\format{An object of class \code{modelorg}}

\references{
Orth, J. D., Conrad, T. M., Na, J., Lerman, J. A., Nam, H., Feist, A. M., & Palsson, B. O. (2011).
 A comprehensive genome-scale reconstruction of Escherichia coli metabolism—2011. 
 \emph{Molecular systems biology}, 7(1).
 }
\keyword{datasets}
\keyword{iJO1366}