\name{iML1515}
\docType{data}
\alias{iML1515}

\encoding{utf8}

\title{Escherichia coli Metabolic Model iML1515}

\description{
  The dataset is a genome scale metabolic network of the \emph{E. coli}. 
  It consists of 2381 internal reactions, 331 exchange reactions and a biomass objective function.
  It contains 1515 genes.
}

\usage{data(iML1515)}

\format{An object of class \code{modelorg}}

\references{
Monk, Jonathan M., et al. "iML1515, a knowledgebase that computes Escherichia coli traits." 
Nature biotechnology 35.10 (2017): 904. }
\keyword{datasets}
\keyword{iML1515}