\name{iMM904}
\docType{data}
\alias{iMM904}

\encoding{utf8}

\title{Saccharomyces cerevisiae Metabolic Model}

\description{
  The dataset is a genome scale metabolic network of the \emph{Saccharomyces cerevisiae}. 
  It consists of  1412 internal reactions, 164 exchange
  reactions and a biomass objective function.
}

\usage{data(iMM904)}

\format{An object of class \code{modelorg}}

\references{
  Mo ML, Palsson BO, Herrgard MJ: Connecting extracellular metabolomic
measurements to intracellular flux states in yeast. BMC Syst Biol 2009,3:37.
}

\keyword{datasets}
