\name{kcat}
\docType{data}
\alias{kcat}

\encoding{utf8}

\title{Escherichia coli KCAT values used in MOMENT method}

\description{
  The dataset is a list of kcat values used in method MOMENT. Values are in unit 1/S.
}

\usage{data(kcat)}

\format{A data frame with three columns}

\references{Adadi, R., Volkmer, B., Milo, R., Heinemann, M., & Shlomi, T. (2012). 
Prediction of Microbial Growth Rate versus Biomass Yield by a Metabolic Network 
with Kinetic Parameters, 8(7). doi:10.1371/journal.pcbi.1002575}

\keyword{datasets}
\keyword{kcat}