\name{mw}
\docType{data}
\alias{mw}

\encoding{utf8}

\title{Escherichia coli molecular weight values used in MOMENT method}

\description{
  The dataset is a list of molecular weights values used in method MOMENT. 
  Values are in unit g/mol.
}

\usage{data(mw)}

\format{A data frame with two columns}

\references{Adadi, R., Volkmer, B., Milo, R., Heinemann, M., & Shlomi, T. (2012). 
Prediction of Microbial Growth Rate versus Biomass Yield by a Metabolic Network 
with Kinetic Parameters, 8(7). doi:10.1371/journal.pcbi.1002575}

\keyword{datasets}
\keyword{mw}