\name{mw_iML1515}
\docType{data}
\alias{mw_iML1515}

\encoding{utf8}

\title{SEColi metabolic genes molecular weight values used in MOMENT/ccFBA method}

\description{
  The dataset is a list of molecular weights values used in method in application of MOMENT method to
  EColi. 
  Values are in unit g/mol.
  Calculated from the EColi genome sequence available at NCBI using \code{calc_MW}.
}

\usage{data(mw_iML1515)}

\format{A data frame with two columns}

\references{
Adadi, R., Volkmer, B., Milo, R., Heinemann, M., & Shlomi, T. (2012). 
Prediction of Microbial Growth Rate versus Biomass Yield by a Metabolic Network 
with Kinetic Parameters, 8(7). doi:10.1371/journal.pcbi.1002575
}

\keyword{datasets}
\keyword{mw_iML1515}