\name{yst_kcat}
\docType{data}
\alias{yst_kcat}

\encoding{utf8}

\title{Saccharomyces cerevisiae KCAT values used in MOMENT method}

\description{
  The dataset is a list of yst_kcat values used in method MOMENT. Values are in unit 1/S.
}

\usage{data(yst_kcat)}

\format{A data frame with three columns}

\references{
Richter C. Kosten und Effizienz von Enzymen als zusaetzliche Bedingung fuer die Flussverteilung in metabolischen Netzwerken 
[Diploma Thesis]. 
Berlin: Humboldt University; 2011.

Adadi, R., Volkmer, B., Milo, R., Heinemann, M., & Shlomi, T. (2012). 
Prediction of Microbial Growth Rate versus Biomass Yield by a Metabolic Network 
with Kinetic Parameters, 8(7). doi:10.1371/journal.pcbi.1002575
}

\keyword{datasets}
\keyword{yst_kcat}