\name{yst_mw}
\docType{data}
\alias{yst_mw}

\encoding{utf8}

\title{Saccharomyces cerevisiae molecular weight values used in MOMENT method}

\description{
  The dataset is a list of molecular weights values used in method in application of MOMENT method to
  Saccharomyces cerevisiae. 
  Values are in unit g/mol.
  Calculated from the yeast genome sequence available at NCBI (Saccharomyces cerevisiae S288c) using \code{calc_MW}.
}

\usage{data(yst_mw)}

\format{A data frame with two columns}

\references{
Adadi, R., Volkmer, B., Milo, R., Heinemann, M., & Shlomi, T. (2012). 
Prediction of Microbial Growth Rate versus Biomass Yield by a Metabolic Network 
with Kinetic Parameters, 8(7). doi:10.1371/journal.pcbi.1002575
}

\keyword{datasets}
\keyword{yst_mw}